refactor(perf-test): Extract `runSuite` into `perf/perf_test.js`
